+++ src/libs/conduit/conduit_utils.cpp
@@ -1541,16 +1541,16 @@ inline unsigned int Hash(const unsigned char *k, unsigned int length, unsigned i
 
    switch(len)
    {
-      case 11: c+=((unsigned int)k[10]<<24);
-      case 10: c+=((unsigned int)k[9]<<16);
-      case 9 : c+=((unsigned int)k[8]<<8);
-      case 8 : b+=((unsigned int)k[7]<<24);
-      case 7 : b+=((unsigned int)k[6]<<16);
-      case 6 : b+=((unsigned int)k[5]<<8);
-      case 5 : b+=k[4];
-      case 4 : a+=((unsigned int)k[3]<<24);
-      case 3 : a+=((unsigned int)k[2]<<16);
-      case 2 : a+=((unsigned int)k[1]<<8);
+      case 11: c+=((unsigned int)k[10]<<24); /* FALLTHROUGH */
+      case 10: c+=((unsigned int)k[9]<<16); /* FALLTHROUGH */
+      case 9 : c+=((unsigned int)k[8]<<8); /* FALLTHROUGH */
+      case 8 : b+=((unsigned int)k[7]<<24); /* FALLTHROUGH */
+      case 7 : b+=((unsigned int)k[6]<<16); /* FALLTHROUGH */
+      case 6 : b+=((unsigned int)k[5]<<8); /* FALLTHROUGH */
+      case 5 : b+=k[4]; /* FALLTHROUGH */
+      case 4 : a+=((unsigned int)k[3]<<24); /* FALLTHROUGH */
+      case 3 : a+=((unsigned int)k[2]<<16); /* FALLTHROUGH */
+      case 2 : a+=((unsigned int)k[1]<<8); /* FALLTHROUGH */
       case 1 : a+=k[0];
    }